fix(update): move translates and styles from hooks to component
