+++ src/screens/CheckEmail/hooks/useCheckEmail.tsx
@@ -1,12 +1,11 @@
 import {useState} from 'react';
-import {useRequestLoading, useTranslation} from 'core/hooks';
+import {useRequestLoading} from 'core/hooks';
 import {useDispatch} from 'react-redux';
 import {useNavigation} from '@react-navigation/native';
 import {checkUserEmailRequest} from 'core/reducers';
 import {CheckEmailScreenNavigationProps} from '../types';
 
 export const useCheckEmail = () => {
-  const {t} = useTranslation('authentification');
   const navigation = useNavigation<CheckEmailScreenNavigationProps>();
   const dispatch = useDispatch();
   const [email, setEmail] = useState('');
@@ -15,7 +14,6 @@ export const useCheckEmail = () => {
   const {loading} = useRequestLoading(checkUserEmailRequest);
 
   return {
-    t,
     navigation,
     email,
     setIsEmailCorrect,
