+++ src/screens/EmailValidation/EmailValidation.tsx
@@ -1,25 +1,26 @@
 import React from 'react';
 
 import {confirmSignUpRequest} from 'core/reducers';
-import {useOnRequestSuccess} from 'core/hooks';
+import {useOnRequestSuccess, useTranslation} from 'core/hooks';
 import {AuthForm} from 'organisms';
 import {OneTimeCode} from 'atoms';
 import {useEmailValidation} from './hooks';
 
 export const EmailValidation = () => {
+  const {t} = useTranslation('authentification');
   const {
-    t,
     navigation,
     isSignUp,
     email,
     onConfirmSignUp,
-    buttonText,
     isCodeFull,
     loading,
     onResendSignUpCodetoEmail,
     onResendRestorePasswordCodetoEmail,
     getEmailCode,
   } = useEmailValidation();
+
+  const buttonText = t('ready').toUpperCase();
 
   useOnRequestSuccess(confirmSignUpRequest, () => {
     navigation.getParent()?.goBack();
