+++ src/screens/Home/Home.tsx
@@ -1,42 +1,51 @@
 import React, {useEffect, useCallback} from 'react';
-import {RefreshPageReminder, SuspenseView} from 'atoms';
+import {RefreshPageReminder, SuspenseView, useToast} from 'atoms';
 import {HomeSectionBar} from 'organisms';
 import {FlatList, InteractionManager, RefreshControl, View} from 'react-native';
 
 import {getHomeDataUpdateAvailableRequest} from 'core/reducers';
 
-import {useAppState} from 'core/hooks';
+import {
+  useAppState,
+  useColorScheme,
+  useThemeStyles,
+  useTranslation,
+} from 'core/hooks';
 import {COLORS} from 'assets';
 import {useFocusEffect} from '@react-navigation/core';
 import {ErrorToast} from '../../components/molecules';
 import {screenOptions} from './screenOptions';
 import {useHome} from './hooks';
+import {themeStyles} from './styles';
+import {useSelector} from 'react-redux';
+import {useIsFocused} from '@react-navigation/native';
+import {selectHomeData, selectIsUpdatesAvailable} from 'core/selectors';
 
 export const Home = () => {
+  const {t} = useTranslation('home');
+  const styles = useThemeStyles(themeStyles);
   const {
-    t,
     updateError,
-    showToast,
     checkUpdates,
     dispatch,
     loading,
-    homeData,
     error,
     listRef,
     getInitialData,
-    styles,
-    theme,
     refreshing,
-    isFocused,
     getData,
     navigateToObjectDetails,
     onCategoryPress,
     onAllObjectsPress,
     navigateToCategoriesList,
     sendIsFavoriteChangedEvent,
-    isUpdatesAvailable,
-    ref,
   } = useHome();
+
+  const theme = useColorScheme();
+  const homeData = useSelector(selectHomeData);
+  const isUpdatesAvailable = useSelector(selectIsUpdatesAvailable);
+  const isFocused = useIsFocused();
+  const {ref, show: showToast} = useToast();
 
   useEffect(() => {
     if (updateError) {
