+++ src/screens/ObjectDetailsMap/hooks/useObjectDetailsMap.tsx
@@ -1,26 +1,18 @@
 import {useCallback, useMemo, useRef} from 'react';
 import MapBox, {RegionPayload} from '@react-native-mapbox-gl/maps';
-import {
-  createMarkerFromDetailsObject,
-  selectIsDirectionShowed,
-  selectMapDirection,
-} from 'core/selectors';
+import {selectIsDirectionShowed} from 'core/selectors';
 import {useDispatch, useSelector} from 'react-redux';
 import {
   useBottomMenu,
-  useColorScheme,
   useFocusToUserLocation,
   useObject,
-  useObjectBelongsToSubtitle,
   useRequestLoading,
-  useThemeStyles,
   useTranslation,
   useTransformedData,
 } from 'core/hooks';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 import {IBounds, IObject} from 'core/types';
 import {showLocation} from 'react-native-map-link';
-import {themeLayerStyles} from '../styles';
 import {mapService} from 'services/MapService';
 import {showObjectDetailsMapDirectionRequest} from 'core/reducers';
 
@@ -38,35 +30,30 @@ import {useRoute} from '@react-navigation/native';
 import {ObjectDetailsMapScreenRouteProps} from '../types';
 
 export const useObjectDetailsMap = () => {
-  const layersStyles = useThemeStyles(themeLayerStyles, {
-    disableStyleSheet: true,
-  });
-  const theme = useColorScheme();
   const {t} = useTranslation('objectDetails');
-  //   const {objectId} = route.params;
   const {
     params: {objectId},
   } = useRoute<ObjectDetailsMapScreenRouteProps>();
-  const camera = useRef<MapBox.Camera>(null);
-  const map = useRef<MapBox.MapView>(null);
 
   const {openMenu, closeMenu, ...menuProps} = useBottomMenu();
 
-  const {bottom, top} = useSafeAreaInsets();
   const data = useObject(objectId);
   const {getObject} = useTransformedData();
-  const dataShapeSource = useMemo(
-    () => (data ? createMarkerFromDetailsObject(data) : data),
-    [data],
-  );
 
-  const belongsToSubtitle = useObjectBelongsToSubtitle(
-    data?.belongsTo?.[0]?.objects,
-  );
+  const map = useRef<MapBox.MapView>(null);
+  const camera = useRef<MapBox.Camera>(null);
 
   const dispatch = useDispatch();
+
+  const centerCoordinate = useMemo(() => {
+    if (data) {
+      return [data.location?.lon!, data.location?.lat!];
+    }
+
+    return null;
+  }, [data]);
+
   const isDirectionShowed = useSelector(selectIsDirectionShowed);
-  const direction = useSelector(selectMapDirection);
 
   const {
     focusToUserLocation,
@@ -76,34 +63,6 @@ export const useObjectDetailsMap = () => {
     userLocation,
     ...userLocationProps
   } = useFocusToUserLocation(camera);
-
-  const bounds = useMemo(() => {
-    if (data) {
-      const paddings = {
-        bottom: 169 + bottom,
-        top: 30 + top,
-      };
-
-      if (data.area) {
-        return mapService.getBoundsFromGeoJSON(data.area, paddings);
-      }
-
-      if (data.routes) {
-        return mapService.getBoundsFromGeoJSON(data.routes, paddings);
-      }
-
-      return null;
-    }
-    return null;
-  }, [bottom, data, top]);
-
-  const centerCoordinate = useMemo(() => {
-    if (data) {
-      return [data.location?.lon!, data.location?.lat!];
-    }
-
-    return null;
-  }, [data]);
 
   const {loading} = useRequestLoading(showObjectDetailsMapDirectionRequest);
 
@@ -136,6 +95,28 @@ export const useObjectDetailsMap = () => {
     },
     [centerCoordinate, dispatch, getUserLocation, isDirectionShowed, t],
   );
+
+  const {bottom, top} = useSafeAreaInsets();
+
+  const bounds = useMemo(() => {
+    if (data) {
+      const paddings = {
+        bottom: 169 + bottom,
+        top: 30 + top,
+      };
+
+      if (data.area) {
+        return mapService.getBoundsFromGeoJSON(data.area, paddings);
+      }
+
+      if (data.routes) {
+        return mapService.getBoundsFromGeoJSON(data.routes, paddings);
+      }
+
+      return null;
+    }
+    return null;
+  }, [bottom, data, top]);
 
   const boundsToArea = useCallback(() => {
     if (bounds) {
@@ -212,26 +193,20 @@ export const useObjectDetailsMap = () => {
   }, [closeMenu]);
 
   return {
-    theme,
-    direction,
     bottom,
     top,
     camera,
     data,
     openMenu,
     dispatch,
-    centerCoordinate,
     unfocusUserLocation,
     onMarkerPress,
     bounds,
     map,
     userLocationProps,
-    layersStyles,
-    dataShapeSource,
     menuProps,
     onShowLocationPress,
     isUserLocationFocused,
-    belongsToSubtitle,
     onMenuButtonPress,
     loading,
     isDirectionShowed,
