+++ src/screens/Profile/Profile.tsx
@@ -3,16 +3,22 @@ import {View} from 'react-native';
 import {GeneralListItem} from 'molecules';
 import {Icon, SettingsSectionTitle} from 'atoms';
 import {useProfile} from './hooks';
+import {useTranslation} from 'react-i18next';
+import {useThemeStyles} from 'core/hooks';
+import {themeStyles} from './styles';
+import {useSelector} from 'react-redux';
+import {selectUserEmail} from 'core/selectors';
 
 export const Profile = () => {
+  const {t} = useTranslation('profile');
+  const styles = useThemeStyles(themeStyles);
   const {
-    t,
-    styles,
+    isAuthorized,
     onAuthorisationItemPress,
     navigateToProfileSettingsTheme,
-    isAuthorized,
-    userEmail,
   } = useProfile();
+
+  const userEmail = useSelector(selectUserEmail);
 
   return (
     <View style={styles.container}>
