+++ src/screens/Profile/hooks/useProfile.tsx
@@ -2,17 +2,12 @@ import {useCallback} from 'react';
 
 import {useNavigation} from '@react-navigation/native';
 import {ProfileScreenNavigationProps} from '../types';
-import {useThemeStyles, useTranslation} from 'core/hooks';
-import {themeStyles} from '../styles';
 import {useSelector} from 'react-redux';
-import {selectUserAuthorized, selectUserEmail} from 'core/selectors';
+import {selectUserAuthorized} from 'core/selectors';
 
 export const useProfile = () => {
-  const styles = useThemeStyles(themeStyles);
   const navigation = useNavigation<ProfileScreenNavigationProps>();
-  const {t} = useTranslation('profile');
   const isAuthorized = useSelector(selectUserAuthorized);
-  const userEmail = useSelector(selectUserEmail);
 
   const onAuthorisationItemPress = useCallback(() => {
     if (isAuthorized) {
@@ -29,11 +24,8 @@ export const useProfile = () => {
   }, [navigation]);
 
   return {
-    t,
-    styles,
+    isAuthorized,
     onAuthorisationItemPress,
     navigateToProfileSettingsTheme,
-    isAuthorized,
-    userEmail,
   };
 };
