+++ src/screens/SignUpForm/SignUpForm.tsx
@@ -1,23 +1,27 @@
 import React from 'react';
-import {useOnRequestSuccess} from 'core/hooks';
+import {
+  useOnRequestSuccess,
+  useTogglePasswordVisibility,
+  useTranslation,
+} from 'core/hooks';
 import {FormInput} from 'atoms';
 import {signUpRequest} from 'core/reducers';
 import {AuthForm} from 'organisms';
 import {useSignUpForm} from './hooks';
 
 export const SignUpForm = () => {
+  const {t} = useTranslation('authentification');
   const {
-    t,
     loading,
     email,
     password,
     setPassword,
-    passwordVisibility,
-    rightIcon,
-    handlePasswordVisibility,
     navigateToEmailValidation,
     signUp,
   } = useSignUpForm();
+
+  const {passwordVisibility, rightIcon, handlePasswordVisibility} =
+    useTogglePasswordVisibility('eye');
 
   useOnRequestSuccess(signUpRequest, navigateToEmailValidation);
