+++ src/screens/Splash/hooks/useSplash.tsx
@@ -1,9 +1,7 @@
 import {useCallback} from 'react';
 import {
-  useAnimatedStyle,
   useSharedValue,
   withTiming,
-  interpolate,
   Easing,
   runOnJS,
 } from 'react-native-reanimated';
@@ -12,32 +10,6 @@ import RNBootSplash from 'react-native-bootsplash';
 export const useSplash = () => {
   const opacity = useSharedValue(1);
   const animatedValue = useSharedValue(0);
-
-  const imageAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      transform: [
-        {translateX: interpolate(animatedValue.value, [0, 1], [0, -90])},
-      ],
-    };
-  });
-
-  const textAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      opacity: animatedValue.value,
-      transform: [
-        {scale: animatedValue.value},
-        {
-          translateX: interpolate(animatedValue.value, [0, 1], [0, 35]),
-        },
-      ],
-    };
-  });
-
-  const containerAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      opacity: opacity.value,
-    };
-  });
 
   const animateIOS = useCallback(
     onAnimationEnd => {
@@ -96,10 +68,9 @@ export const useSplash = () => {
   );
 
   return {
+    opacity,
+    animatedValue,
     animateIOS,
     animateAndroid,
-    containerAnimatedStyle,
-    imageAnimatedStyle,
-    textAnimatedStyle,
   };
 };
