+++ src/storage/types.ts
@@ -7,8 +7,4 @@ export type StorageOptions = {
 
 export type StorageResponse = string | null | undefined;
 
-export type StorageKeys =
-  | 'allAppDataVersion'
-  | 'appPrevLocale'
-  | 'etag'
-  | 'appPrevTheme';
+export type StorageKeys = 'allAppDataVersion' | 'appPrevLocale' | 'etag';
